docs: revert relocation
