+++ packages/odyssey-storybook/src/components/Box/Box.mdx
@@ -14,7 +14,7 @@ A simple `<Button>` component might look like this:
 </Box>
 ```
 
-[View the source on GitHub](https://github.com/okta/odyssey/blob/develop/packages/odyssey-react/src/components/Box/Box.tsx)
+[View the source on GitHub](https://github.com/okta/odyssey/tree/master/packages/odyssey-react/src/components/Box)
 
 ## Props
