+++ packages/odyssey-storybook/src/components/Text/TextStyles.stories.tsx
@@ -17,7 +17,7 @@ import type { TextProps } from "@okta/odyssey-react";
 import { Text as Source } from "../../../../odyssey-react/src";
 
 export default {
-  title: `Utilities/Text/Styles`,
+  title: `Components/Text/Styles`,
   component: Source,
   argTypes: {
     children: {
