--- packages/odyssey-storybook/src/utilities/Utilities.stories.mdx
@@ -1,155 +1,0 @@
-import { Meta } from "@storybook/addon-docs/blocks";
-
-<Meta title="Utilities/Utilities" />
-
-# Utilities
-
-## Box VS Text
-
-In production docs, these sections will live with their appropriate utility. I've co-located them here, so that we can align internally.
-
-### Box
-
-### Text
-
-**WIP API NOTE:** Would anyone be opposed to renaming this component to `<StyledElement>` or similar?
-
-`<Text />` is a convenience utility built with `<Box />`. It can provide its own hooks for customization and theme-able tokens as necessary.
-
-#### Styled element insertion
-
-`<Text />` enables authors and components to insert styled HTML5 elements without the need for a global CSS sheet.
-
-```jsx
-We <Text as="em">really</Text> need to get off this planet.
-```
-
-A more complicated example might look like this:
-
-```jsx
-<Text as="blockquote">
-  <Text as="p">
-    The important thing is not to stop questioning. Curiosity has its own reason
-    for existence.
-  </Text>
-  <footer>
-    Albert Einstein,
-    <Text as="cite">
-      Old Man's Advice to Youth: "Never Lose a Holy Curiosity," LIFE magazine (2
-      May 1955) statement to William Miller, p. 64.
-    </Text>
-  </footer>
-</Text>
-```
-
-####
-
-**WIP API NOTE:** This next section is an additional use-case for `<Text />`.
-
-In addition, `<Text />` provides a simpler API with indirect descendant selectors for natural HTML5 composition or styling dynamic content (e.g. from markdown).
-
-```jsx
-<Text>
-  <blockquote>
-    <p>
-      The important thing is not to stop questioning. Curiosity has its own
-      reason for existence.
-    </p>
-    <footer>
-      Albert Einstein,
-      <cite>
-        Old Man's Advice to Youth: "Never Lose a Holy Curiosity," LIFE magazine
-        (2 May 1955) statement to William Miller, p. 64.
-      </cite>
-    </footer>
-  </blockquote>
-</Text>
-```
-
-## Current Cases
-
-Currently, we have a few open PRs that stand as practical examples for application of these utilities. Below, I've outlined my current understanding of what each component should look like, given these utilities.
-
-### Field
-
-```jsx
-// Field
-
-<Box as={tag} className={styles.root}>
-  <Field.Label
-    inputId={inputId}
-    required={required}
-    optionalLabel={optionalLabel}
-    labelHidden={labelHidden}
-    as={TagLabel}
-  >
-    {label}
-  </Field.Label>
-  {hint && <Field.Hint id={inputId}>{hint}</Field.Hint>} {children} {error &&
-  <Field.Error id={inputId}>{error}</Field.Error>}
-</Box>
-
-// Label //// In the case of these label elements, we can treat our Field
-styling as the HTML5 default, removing the need for component-specific styling.
-
-<Text as={labelTag} htmlFor={inputId}>
-  {children} {!required && optionalLabel && (
-  <span className={styles.optionalLabel} children={optionalLabel} />
-  )}
-</Text>
-
-// Hidden Label
-
-<ScreenReaderText>
-  <Text as={labelTag} htmlFor={inputId}>
-    {children} {!required && optionalLabel && (
-    <span className={styles.optionalLabel} children={optionalLabel} />
-    )}
-  </Text>
-</ScreenReaderText>
-
-// Hint
-
-<Box
-  as="p"
-  className={styles.hint}
-  id={`${id}-hint`}
-  children={children}
-/>
-
-// Error
-
-<Box as="p" className={styles.error} id={`${id}-error`}>{children} /> </Box>
-```
-
-### Heading
-
-Much like `<Text />` is a convenience wrapper on top of `<Box />`, `<Heading />` is built on `<Text />`.
-
-```jsx
-// Heading
-
-<Text as={tag} className={componentClass}>
-  {children}
-</Text>
-```
-
-My assumption here is that all Headings have parity with our desired HTML5 defaults. The `className` would be used to apply a different level of visual hierarchy to a tag (an `h1` with the styles of an `h3`).
-
-If a component requires heading styling that deviates farther than that, those styles should live with the component until they become a pattern.
-
-### List
-
-```jsx
-// List
-
-<Box as={tag}> { children } </Box>
-
-//// `tag` here would be ul, ol, dl // ListItem
-
-<Text as={childTag}> { children } </Text>
-```
-
-If `<List>` were only intended to be used in the fashion of plain HTML5, `<Text>` could probably be used throughout. e.g. `<Text as="ul">`. However, we've seen that `<List>` is likely to have more complicated functionality.
-
-Alternately, using `<Text as="ul">` could provide default styling while `<List>` comes without assumed margins and padding.
